Restructure browser test directory
