+++ browserTests/generic/browserTest.js
@@ -3,8 +3,8 @@ import { Selector, ClientFunction } from 'testcafe';
 
 import { waitForReact, ReactSelector } from 'testcafe-react-selectors';
 
-import { paletteDefault, paletteDark } from '../src/themes'
-import config from './config';
+import { paletteDefault, paletteDark } from '../../src/themes'
+import config from '../config';
 const { server } = config;
 
 const siteRoot = `http://${server.address}:${server.port}`;
