+++ browserTests/generic/mapTest.js
@@ -1,5 +1,5 @@
 import { waitForReact, ReactSelector } from 'testcafe-react-selectors';
-import config from './config';
+import config from '../config';
 import { Selector } from 'testcafe';
 
 /* eslint-disable */
