+++ package.json
@@ -59,7 +59,7 @@
     "eject": "react-scripts eject",
     "dev": "NODE_ENV=development BABEL_ENV=development webpack --watch --progress & nodemon dist",
     "browserTest": "node browserTests/browserTestRunner.js",
-    "browserTestCLI": "testcafe chrome:headless browserTests/*Test.js -q attemptLimit=3,successThreshold=1 --app \"node dist\"",
+    "browserTestCLI": "testcafe chrome:headless browserTests/*/*Test.js -q attemptLimit=3,successThreshold=1 --app \"node dist\"",
     "production": "NODE_ENV=production BABEL_ENV=production webpack --progress --mode production && node dist"
   },
   "eslintConfig": {
